requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
arabic-reshaper==2.1.3
    # via xhtml2pdf
asgiref==3.5.2
    # via django
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   pyhanko
    #   pyhanko-certvalidator
attrs==22.1.0
    # via pytest
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via pyhanko
coverage==6.4.4
    # via -r requirements.in
crispy-bootstrap5==0.6
    # via -r requirements.in
cryptography==37.0.4
    # via
    #   pyhanko
    #   pyhanko-certvalidator
cssselect2==0.6.0
    # via svglib
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
django==4.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-debug-toolbar
    #   django-extensions
    #   django-import-export
django-crispy-forms==1.14.0
    # via crispy-bootstrap5
django-debug-toolbar==3.6.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
django-extensions==3.2.0
    # via -r requirements.in
django-import-export==2.8.0
    # via -r requirements.in
django-simple-history==3.1.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
future==0.18.2
    # via arabic-reshaper
html5lib==1.1
    # via xhtml2pdf
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
lxml==4.9.1
    # via svglib
markuppy==1.14
    # via tablib
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
oscrypto==1.3.0
    # via pyhanko-certvalidator
packaging==21.3
    # via
    #   pytest
    #   pytest-sugar
pillow==9.2.0
    # via
    #   reportlab
    #   xhtml2pdf
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pyhanko==0.13.2
    # via xhtml2pdf
pyhanko-certvalidator==0.19.5
    # via
    #   pyhanko
    #   xhtml2pdf
pyparsing==3.0.9
    # via packaging
pypdf3==1.0.6
    # via xhtml2pdf
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-sugar
pytest-django==4.5.2
    # via -r requirements.in
pytest-sugar==0.9.5
    # via -r requirements.in
python-bidi==0.4.2
    # via xhtml2pdf
pytz==2022.2.1
    # via pyhanko
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   pyhanko
    #   tablib
qrcode==7.3.1
    # via pyhanko
reportlab==3.6.11
    # via
    #   svglib
    #   xhtml2pdf
requests==2.28.1
    # via
    #   pyhanko
    #   pyhanko-certvalidator
six==1.16.0
    # via
    #   html5lib
    #   python-bidi
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
svglib==1.4.1
    # via xhtml2pdf
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
termcolor==1.1.0
    # via pytest-sugar
tinycss2==1.1.1
    # via
    #   cssselect2
    #   svglib
tomli==2.0.1
    # via pytest
tqdm==4.64.0
    # via pypdf3
tzdata==2022.2
    # via pytz-deprecation-shim
tzlocal==4.2
    # via pyhanko
uritools==4.0.0
    # via pyhanko-certvalidator
urllib3==1.26.12
    # via requests
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
xhtml2pdf==0.2.8
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
